requirements.txt
Django>=4.0
django-allauth
